+++ include/gdk.h
@@ -458,6 +458,17 @@ GDK_API int GA_set_pin(
     struct GA_session* session, const char* mnemonic, const char* pin, const char* device_id, GA_json** pin_data);
 
 /**
+ * Encrypt json with server provided key protected by a PIN.
+ *
+ * :param session: The session to use.
+ * :param details: The :ref:`encrypt-with-pin-details` to encrypt.
+ * :param call: Destination for the resulting GA_auth_handler to complete the action.
+ *|     The call handlers result is :ref:`encrypt-with-pin-result` which the caller should persist.
+ *|     Returned GA_auth_handler should be freed using `GA_destroy_auth_handler`.
+ */
+GDK_API int GA_encrypt_with_pin(struct GA_session* session, const GA_json* details, struct GA_auth_handler** call);
+
+/**
  * Disable all PIN logins previously set.
  *
  * After calling this method, the user will not be able to login with PIN