+++ src/session.hpp
@@ -53,6 +53,7 @@ namespace sdk {
         nlohmann::json get_twofactor_config(bool reset_cached = false);
 
         nlohmann::json set_pin(const std::string& mnemonic, const std::string& pin, const std::string& device_id);
+        nlohmann::json encrypt_with_pin(const nlohmann::json& details);
         void disable_all_pin_logins();
 
         nlohmann::json get_unspent_outputs_for_private_key(