+++ src/swift/GreenAddress/Sources/GreenAddress/GreenAddress.swift
@@ -362,6 +362,10 @@ public class Session {
         return try convertOpaqueJsonToDict(o: result!)
     }
 
+    public func encryptWithPin(details: [String: Any]) throws -> TwoFactorCall {
+        return try jsonFuncToCallHandlerWrapper(input: details, fun: GA_encrypt_with_pin)
+    }
+
     public func disableAllPinLogins() throws -> Void {
         try callWrapper(fun: GA_disable_all_pin_logins(session))
     }