+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Singlesig: Implement GA_get_previous_addresses
 - Singlesig: Allow fetching internal addresses in GA_get_receive_address using "is_internal"=true
 - Added new function GA_get_credentials, to replace GA_get_mnemonic_passphrase.
+- Added new function GA_encrypt_with_pin, to replace GA_set_pin.
 
 ### Changed